server/requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
amqp==5.0.6
ansible==5.0.1
ansible-core==2.12.0
asgiref==3.4.1
async-timeout==4.0.1
attrs==21.2.0
base58==2.1.1
billiard==3.6.4.0
bitarray==1.2.2
celery==5.2.1
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.8
click==8.0.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
cryptography==36.0.0
cytoolz==0.11.2
Django==3.2.9
django-celery-beat==2.2.1
django-environ==0.8.1
django-ethereum-events==4.1.1
django-solo==1.2.0
django-timezone-field==4.2.1
djangorestframework==3.12.4
eth-abi==2.1.1
eth-account==0.5.6
eth-hash==0.3.2
eth-keyfile==0.5.1
eth-keys==0.3.3
eth-rlp==0.2.1
eth-typing==2.2.2
eth-utils==1.10.0
frozenlist==1.2.0
gunicorn==20.1.0
hexbytes==0.2.2
idna==3.3
ipfshttpclient==0.8.0a2
Jinja2==3.0.3
jsonschema==3.2.0
kombu==5.2.2
lru-dict==1.1.7
MarkupSafe==2.0.1
multiaddr==0.0.9
multidict==5.2.0
netaddr==0.8.0
packaging==21.3
parsimonious==0.8.1
prompt-toolkit==3.0.23
protobuf==3.19.1
pycparser==2.21
pycryptodome==3.11.0
pyparsing==3.0.6
pyrsistent==0.18.0
python-crontab==2.6.0
python-dateutil==2.8.2
pytz==2021.3
PyYAML==6.0
redis==3.5.3
requests==2.26.0
resolvelib==0.5.4
rlp==2.0.1
six==1.16.0
sqlparse==0.4.2
toolz==0.11.2
typing_extensions==4.0.0
urllib3==1.26.7
varint==1.0.2
vine==5.0.0
wcwidth==0.2.5
web3==5.25.0
websockets==9.1
yarl==1.7.2
